requirements.txt
# Core dependencies
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
passlib[scrypt]==1.7.4
mysql-connector-python==9.2.0
python-dotenv==1.0.1  # Optional, only if you're loading .env config files
pytz==2024.1
stripe 
pytest>=7.0
opencv-python>=4.8
fpdf2>=2.7

# Google API Client dependencies
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
Flask-Mail

# Google Generative AI
google-generativeai==0.3.2

# PDF generation
fpdf==1.7.2
pdf2image==1.16.3

# CLI and Terminal Formatting
termcolor==2.4.0
prettytable==3.10.0
